backend/requirements.txt
# FastAPI & Web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1

# File processing
openpyxl==3.1.2
pandas==2.2.0
pdfplumber==0.11.8

# OpenAI
openai==1.12.0

# HTTP & APIs
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0

# Notion
notion-client==2.2.1

# AWS S3 compatible storage (for Cloudflare R2)
boto3==1.34.34

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0

# Utils
python-dotenv==1.0.1
pytz==2024.1
Pillow==10.2.0
